+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * **BUGFIX:** Fixed the shell scripts in the Getting Started page to be compatible with bash.
 * **BUGFIX:** Fixed a bug where test payloads containing results with a null category would not be processed.
+* **BUGFIX:** Fixed a bug where anonymous users would see private organizations in the list of accessible organizations.